+++ app/src/main/java/com/github/dfa/diaspora_android/activity/MainActivity.java
@@ -54,6 +54,7 @@ import android.widget.EditText;
 import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
+import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -68,18 +69,25 @@ import com.github.dfa.diaspora_android.fragment.HashtagListFragment;
 import com.github.dfa.diaspora_android.fragment.PodSelectionFragment;
 import com.github.dfa.diaspora_android.listener.WebUserProfileChangedListener;
 import com.github.dfa.diaspora_android.receiver.OpenExternalLinkReceiver;
+import com.github.dfa.diaspora_android.util.ProxyHandler;
 import com.github.dfa.diaspora_android.receiver.UpdateTitleReceiver;
 import com.github.dfa.diaspora_android.ui.BadgeDrawable;
+import com.github.dfa.diaspora_android.ui.IntellihideToolbarActivityListener;
 import com.github.dfa.diaspora_android.util.AppLog;
 import com.github.dfa.diaspora_android.util.CustomTabHelpers.CustomTabActivityHelper;
 import com.github.dfa.diaspora_android.util.DiasporaUrlHelper;
+import com.github.dfa.diaspora_android.util.theming.ThemeHelper;
 import com.github.dfa.diaspora_android.util.WebHelper;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import butterknife.OnClick;
 
-public class MainActivity extends AppCompatActivity
-        implements NavigationView.OnNavigationItemSelectedListener, WebUserProfileChangedListener, CustomTabActivityHelper.ConnectionCallback {
+public class MainActivity extends ThemedActivity
+        implements NavigationView.OnNavigationItemSelectedListener,
+        WebUserProfileChangedListener,
+        CustomTabActivityHelper.ConnectionCallback,
+        IntellihideToolbarActivityListener {
 
 
     public static final int REQUEST_CODE_ASK_PERMISSIONS = 123;
@@ -92,7 +100,6 @@ public class MainActivity extends AppCompatActivity
     public static final String ACTION_CHANGE_ACCOUNT = "com.github.dfa.diaspora_android.MainActivity.change_account";
     public static final String ACTION_CLEAR_CACHE = "com.github.dfa.diaspora_android.MainActivity.clear_cache";
     public static final String ACTION_UPDATE_TITLE_FROM_URL = "com.github.dfa.diaspora_android.MainActivity.set_title";
-    public static final String ACTION_RELOAD_ACTIVITY = "com.github.dfa.diaspora_android.MainActivity.reload_activity";
     public static final String URL_MESSAGE = "URL_MESSAGE";
     public static final String EXTRA_URL = "com.github.dfa.diaspora_android.extra_url";
     public static final String CONTENT_HASHTAG = "content://com.github.dfa.diaspora_android.mainactivity/";
@@ -113,6 +120,9 @@ public class MainActivity extends AppCompatActivity
     /**
      * UI Bindings
      */
+    @BindView(R.id.main__appbar)
+    AppBarLayout appBarLayout;
+
     @BindView(R.id.main__topbar)
     Toolbar toolbarTop;
 
@@ -127,6 +137,9 @@ public class MainActivity extends AppCompatActivity
 
     @BindView(R.id.main__navdrawer)
     DrawerLayout navDrawer;
+
+    RelativeLayout navDrawerLayout;
+    LinearLayout navProfilePictureArea;
 
 
     // NavHeader cannot be bound by Butterknife
@@ -160,6 +173,7 @@ public class MainActivity extends AppCompatActivity
         urls = new DiasporaUrlHelper(appSettings);
         customTabActivityHelper = new CustomTabActivityHelper();
         customTabActivityHelper.setConnectionCallback(this);
+        ProxyHandler.getInstance().updateProxySettings(this);
 
         fm = getSupportFragmentManager();
         setupUI();
@@ -169,7 +183,7 @@ public class MainActivity extends AppCompatActivity
             @Override
             public void setTitle(int rId) {
                 CustomFragment top = getTopFragment();
-                if (top != null && top.getFragmentTag().equals(DiasporaStreamFragment.TAG)) {
+                if(top != null && top.getFragmentTag().equals(DiasporaStreamFragment.TAG)) {
                     MainActivity.this.setTitle(rId);
                 }
             }
@@ -177,7 +191,7 @@ public class MainActivity extends AppCompatActivity
             @Override
             public void setTitle(String title) {
                 CustomFragment top = getTopFragment();
-                if (top != null && top.getFragmentTag().equals(DiasporaStreamFragment.TAG)) {
+                if(top != null && top.getFragmentTag().equals(DiasporaStreamFragment.TAG)) {
                     MainActivity.this.setTitle(title);
                 }
             }
@@ -229,19 +243,11 @@ public class MainActivity extends AppCompatActivity
 
         // Load app settings
         setupNavigationSlider();
-
-        if (!appSettings.isIntellihideToolbars()) {
-            AppLog.v(this, "Disable intelligent hiding of toolbars");
-            AppBarLayout.LayoutParams params = (AppBarLayout.LayoutParams) toolbarTop.getLayoutParams();
-            params.setScrollFlags(0);  // clear all scroll flags
-        }
-
         AppLog.v(this, "UI successfully set up");
     }
 
     /**
      * Show DiasporaStreamFragment if necessary and load URL url
-     *
      * @param url URL to load in the DiasporaStreamFragment
      */
     public void openDiasporaUrl(String url) {
@@ -255,13 +261,12 @@ public class MainActivity extends AppCompatActivity
      * Get an instance of the CustomFragment with the tag fragmentTag.
      * If there was no instance so far, create a new one and add it to the FragmentManagers pool.
      * If there is no Fragment with the corresponding Tag, return the top fragment.
-     *
      * @param fragmentTag tag
      * @return corresponding Fragment
      */
     protected CustomFragment getFragment(String fragmentTag) {
         CustomFragment fragment = (CustomFragment) fm.findFragmentByTag(fragmentTag);
-        if (fragment != null) {
+        if(fragment != null) {
             return fragment;
         } else {
             switch (fragmentTag) {
@@ -282,8 +287,8 @@ public class MainActivity extends AppCompatActivity
                     fm.beginTransaction().add(psf, fragmentTag).commit();
                     return psf;
                 default:
-                    AppLog.e(this, "Invalid Fragment Tag: " + fragmentTag
-                            + "\nAdd Fragments Tag to getFragment()'s switch case.");
+                    AppLog.e(this,"Invalid Fragment Tag: "+fragmentTag
+                            +"\nAdd Fragments Tag to getFragment()'s switch case.");
                     return getTopFragment();
             }
         }
@@ -291,13 +296,12 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Show the Fragment fragment in R.id.fragment_container. If the fragment was already visible, do nothing.
-     *
      * @param fragment Fragment to show
      */
     protected void showFragment(CustomFragment fragment) {
         AppLog.v(this, "showFragment()");
         CustomFragment currentTop = (CustomFragment) fm.findFragmentById(R.id.fragment_container);
-        if (currentTop == null || !currentTop.getFragmentTag().equals(fragment.getFragmentTag())) {
+        if(currentTop == null || !currentTop.getFragmentTag().equals(fragment.getFragmentTag())) {
             AppLog.v(this, "Fragment was not visible. Replace it.");
             fm.beginTransaction().addToBackStack(null).replace(R.id.fragment_container, fragment, fragment.getFragmentTag()).commit();
             invalidateOptionsMenu();
@@ -319,9 +323,10 @@ public class MainActivity extends AppCompatActivity
         navView.setNavigationItemSelectedListener(this);
 
         View navHeader = navView.getHeaderView(0);
-        LinearLayout navheaderProfileSection = ButterKnife.findById(navHeader, R.id.nav_profile_picture);
+        navProfilePictureArea = ButterKnife.findById(navHeader, R.id.nav_profile_picture);
+        navDrawerLayout = ButterKnife.findById(navHeader, R.id.nav_drawer);
         //Handle clicks on profile picture
-        navheaderProfileSection.setOnClickListener(new View.OnClickListener() {
+        navProfilePictureArea.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
                 navDrawer.closeDrawer(GravityCompat.START);
@@ -358,11 +363,30 @@ public class MainActivity extends AppCompatActivity
                 }
             }
         }
+        updateNavigationViewEntryVisibilities();
     }
 
+    protected void updateNavigationViewEntryVisibilities() {
+        Menu navMenu = navView.getMenu();
+        navMenu.findItem(R.id.nav_exit).setVisible(appSettings.isVisibleInNavExit());
+        navMenu.findItem(R.id.nav_activities).setVisible(appSettings.isVisibleInNavActivities());
+        navMenu.findItem(R.id.nav_aspects).setVisible(appSettings.isVisibleInNavAspects());
+        navMenu.findItem(R.id.nav_commented).setVisible(appSettings.isVisibleInNavCommented());
+        navMenu.findItem(R.id.nav_followed_tags).setVisible(appSettings.isVisibleInNavFollowed_tags());
+        navMenu.findItem(R.id.nav_about).setVisible(appSettings.isVisibleInNavHelp_license());
+        navMenu.findItem(R.id.nav_liked).setVisible(appSettings.isVisibleInNavLiked());
+        navMenu.findItem(R.id.nav_mentions).setVisible(appSettings.isVisibleInNavMentions());
+        navMenu.findItem(R.id.nav_profile).setVisible(appSettings.isVisibleInNavProfile());
+        navMenu.findItem(R.id.nav_public).setVisible(appSettings.isVisibleInNavPublic_activities());
+
+        // Top bar
+        if (appSettings.getPod() == null) {
+            navMenu.setGroupVisible(navMenu.findItem(R.id.nav_exit).getGroupId(), false);
+        }
+    }
+    
     /**
      * Forward incoming intents to handleIntent()
-     *
      * @param intent incoming
      */
     @Override
@@ -373,7 +397,6 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Handle intents and execute intent specific actions
-     *
      * @param intent intent to get handled
      */
     private void handleIntent(Intent intent) {
@@ -398,7 +421,7 @@ public class MainActivity extends AppCompatActivity
                 return;
             } else {
                 loadUrl = intent.getDataString();
-                AppLog.v(this, "Intent has a delicious URL for us: " + loadUrl);
+                AppLog.v(this, "Intent has a delicious URL for us: "+loadUrl);
             }
         } else if (ACTION_CHANGE_ACCOUNT.equals(action)) {
             AppLog.v(this, "Reset pod data and  show PodSelectionFragment");
@@ -408,10 +431,6 @@ public class MainActivity extends AppCompatActivity
         } else if (ACTION_CLEAR_CACHE.equals(action)) {
             AppLog.v(this, "Clear WebView cache");
             ((DiasporaStreamFragment) getFragment(DiasporaStreamFragment.TAG)).getWebView().clearCache(true);
-        } else if (ACTION_RELOAD_ACTIVITY.equals(action)) {
-            AppLog.v(this, "Recreate activity");
-            recreate();
-            return;
         } else if (Intent.ACTION_SEND.equals(action) && type != null) {
             switch (type) {
                 case "text/plain":
@@ -438,25 +457,23 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Handle activity results
-     *
      * @param requestCode reqCode
-     * @param resultCode  resCode
-     * @param data        data
+     * @param resultCode resCode
+     * @param data data
      */
     @Override
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        AppLog.v(this, "onActivityResult(): " + requestCode);
+        AppLog.v(this, "onActivityResult(): "+requestCode);
         super.onActivityResult(requestCode, resultCode, data);
     }
 
     /**
      * Return the fragment which is currently displayed in R.id.fragment_container
-     *
      * @return top fragment or null if there is none displayed
      */
     private CustomFragment getTopFragment() {
         Fragment top = fm.findFragmentById(R.id.fragment_container);
-        if (top != null) {
+        if(top != null) {
             return (CustomFragment) top;
         }
         return null;
@@ -473,12 +490,12 @@ public class MainActivity extends AppCompatActivity
             return;
         }
         CustomFragment top = getTopFragment();
-        if (top != null) {
+        if(top != null) {
             AppLog.v(this, "Top Fragment is not null");
-            if (!top.onBackPressed()) {
+            if(!top.onBackPressed()) {
                 AppLog.v(this, "Top Fragment.onBackPressed was false");
-                AppLog.v(this, "BackStackEntryCount: " + fm.getBackStackEntryCount());
-                if (fm.getBackStackEntryCount() > 0) {
+                AppLog.v(this, "BackStackEntryCount: "+fm.getBackStackEntryCount());
+                if(fm.getBackStackEntryCount()>0) {
                     fm.popBackStack();
                 } else {
                     snackbarExitApp.show();
@@ -523,12 +540,18 @@ public class MainActivity extends AppCompatActivity
         AppLog.v(this, "Register BroadcastReceivers");
         LocalBroadcastManager.getInstance(this).registerReceiver(brSetTitle, new IntentFilter(ACTION_UPDATE_TITLE_FROM_URL));
         LocalBroadcastManager.getInstance(this).registerReceiver(brOpenExternalLink, new IntentFilter(ACTION_OPEN_EXTERNAL_URL));
+        this.appSettings = getAppSettings();
+        if (appSettings.isIntellihideToolbars()) {
+            this.enableToolbarHiding();
+        } else {
+            this.disableToolbarHiding();
+        }
+        updateNavigationViewEntryVisibilities();
     }
 
     /**
      * Clear and repopulate top and bottom toolbar.
      * Also add menu items of the displayed fragment
-     *
      * @param menu top toolbar
      * @return boolean
      */
@@ -541,9 +564,9 @@ public class MainActivity extends AppCompatActivity
         toolbarBottom.setVisibility(View.VISIBLE);
 
         CustomFragment top = getTopFragment();
-        if (top != null) {
+        if(top != null) {
             //Are we displaying a Fragment other than PodSelectionFragment?
-            if (!top.getFragmentTag().equals(PodSelectionFragment.TAG)) {
+            if(!top.getFragmentTag().equals(PodSelectionFragment.TAG)) {
                 getMenuInflater().inflate(R.menu.main__menu_top, menu);
                 getMenuInflater().inflate(R.menu.main__menu_bottom, toolbarBottom.getMenu());
                 top.onCreateBottomOptionsMenu(toolbarBottom.getMenu(), getMenuInflater());
@@ -559,30 +582,13 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Set the notification and messages counter in the top toolbar
-     *
      * @param menu menu
      * @return boolean
      */
     @Override
     public boolean onPrepareOptionsMenu(Menu menu) {
-        // Navigation slider
-        Menu navMenu = navView.getMenu();
-        navMenu.findItem(R.id.nav_exit).setVisible(appSettings.isVisibleInNavExit());
-        navMenu.findItem(R.id.nav_activities).setVisible(appSettings.isVisibleInNavActivities());
-        navMenu.findItem(R.id.nav_aspects).setVisible(appSettings.isVisibleInNavAspects());
-        navMenu.findItem(R.id.nav_commented).setVisible(appSettings.isVisibleInNavCommented());
-        navMenu.findItem(R.id.nav_followed_tags).setVisible(appSettings.isVisibleInNavFollowed_tags());
-        navMenu.findItem(R.id.nav_about).setVisible(appSettings.isVisibleInNavHelp_license());
-        navMenu.findItem(R.id.nav_liked).setVisible(appSettings.isVisibleInNavLiked());
-        navMenu.findItem(R.id.nav_mentions).setVisible(appSettings.isVisibleInNavMentions());
-        navMenu.findItem(R.id.nav_profile).setVisible(appSettings.isVisibleInNavProfile());
-        navMenu.findItem(R.id.nav_public).setVisible(appSettings.isVisibleInNavPublic_activities());
-
-        // Top bar
         MenuItem item;
-        if (appSettings.getPod() == null) {
-            navMenu.setGroupVisible(navView.getMenu().findItem(R.id.nav_exit).getGroupId(), false);
-        }
+        updateNavigationViewEntryVisibilities();
 
         if ((item = menu.findItem(R.id.action_notifications)) != null) {
             LayerDrawable icon = (LayerDrawable) item.getIcon();
@@ -598,7 +604,6 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Handle clicks on the optionsmenu
-     *
      * @param item item
      * @return boolean
      */
@@ -697,7 +702,6 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Update the profile name in the navigation slider
-     *
      * @param name name
      */
     @Override
@@ -708,7 +712,6 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Update the profile picture in the navigation slider
-     *
      * @param avatarUrl url of the new profile pic
      */
     @Override
@@ -719,7 +722,6 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Handle hashtag clicks. Open the new-post-url and inject the clicked hashtag into the post-editor
-     *
      * @param intent intent
      */
     private void handleHashtag(Intent intent) {
@@ -734,7 +736,6 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Open the new-post-url and inject text that was shared into the app into the post editors text field
-     *
      * @param intent shareTextIntent
      */
     private void handleSendText(Intent intent) {
@@ -793,7 +794,6 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Share an image shared to the app via diaspora
-     *
      * @param intent shareImageIntent
      */
     //TODO: Implement some day
@@ -810,7 +810,6 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Invalidate the top toolbar to update the notification counter
-     *
      * @param notificationCount new notification count
      */
     @Override
@@ -822,7 +821,6 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Invalidate the top toolbar to update the unread messages counter
-     *
      * @param unreadMessageCount new unread messages count
      */
     @Override
@@ -834,8 +832,8 @@ public class MainActivity extends AppCompatActivity
 
     @Override
     public void onCustomTabsConnected() {
-        if (customTabsSession == null) {
-            AppLog.i(this, "CustomTabs warmup: " + customTabActivityHelper.warmup(0));
+        if(customTabsSession == null) {
+            AppLog.i(this, "CustomTabs warmup: "+customTabActivityHelper.warmup(0));
             customTabsSession = customTabActivityHelper.getSession();
         }
     }
@@ -955,9 +953,8 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * React to results of requestPermission
-     *
-     * @param requestCode  resCode
-     * @param permissions  requested permissions
+     * @param requestCode resCode
+     * @param permissions requested permissions
      * @param grantResults granted results
      */
     @Override
@@ -981,7 +978,6 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Return the string that will be shared into the new-post-editor
-     *
      * @return String
      */
     public String getTextToBeShared() {
@@ -990,10 +986,34 @@ public class MainActivity extends AppCompatActivity
 
     /**
      * Set the string that will be shared into the new-post-editor
-     *
      * @param textToBeShared
      */
     public void setTextToBeShared(String textToBeShared) {
         this.textToBeShared = textToBeShared;
+    }
+
+    @Override
+    protected void applyColorToViews() {
+        ThemeHelper.updateToolbarColor(toolbarTop);
+        ThemeHelper.updateActionMenuViewColor(toolbarBottom);
+        navDrawerLayout.setBackgroundColor(appSettings.getPrimaryColor());
+        navProfilePictureArea.setBackgroundColor(appSettings.getPrimaryColor());
+    }
+
+    @Override
+    public void enableToolbarHiding() {
+        AppLog.d(this, "Enable Intellihide");
+        AppBarLayout.LayoutParams params = (AppBarLayout.LayoutParams) toolbarTop.getLayoutParams();
+        //scroll|enterAlways|snap
+        params.setScrollFlags(toolbarDefaultScrollFlags);
+        appBarLayout.setExpanded(true, true);
+    }
+
+    @Override
+    public void disableToolbarHiding() {
+        AppLog.d(this, "Disable Intellihide");
+        AppBarLayout.LayoutParams params = (AppBarLayout.LayoutParams) toolbarTop.getLayoutParams();
+        params.setScrollFlags(0);  // clear all scroll flags
+        appBarLayout.setExpanded(true, true);
     }
 }