+++ app/src/main/java/com/github/dfa/diaspora_android/fragment/BrowserFragment.java
@@ -1,9 +1,24 @@
+/*
+    This file is part of the Diaspora for Android.
+
+    Diaspora for Android is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    Diaspora for Android is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with the Diaspora for Android.
+
+    If not, see <http://www.gnu.org/licenses/>.
+ */
 package com.github.dfa.diaspora_android.fragment;
 
 import android.Manifest;
-import android.app.AlarmManager;
-import android.app.PendingIntent;
-import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
@@ -11,7 +26,6 @@ import android.graphics.Bitmap;
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Environment;
-import android.os.StrictMode;
 import android.support.design.widget.Snackbar;
 import android.support.v7.app.AlertDialog;
 import android.view.LayoutInflater;
@@ -22,16 +36,17 @@ import android.view.ViewGroup;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.widget.ProgressBar;
-import android.widget.Toast;
 
 import com.github.dfa.diaspora_android.App;
 import com.github.dfa.diaspora_android.R;
 import com.github.dfa.diaspora_android.activity.MainActivity;
 import com.github.dfa.diaspora_android.data.AppSettings;
+import com.github.dfa.diaspora_android.util.ProxyHandler;
 import com.github.dfa.diaspora_android.ui.ContextMenuWebView;
+import com.github.dfa.diaspora_android.util.theming.ThemeHelper;
+import com.github.dfa.diaspora_android.util.AppLog;
 import com.github.dfa.diaspora_android.webview.CustomWebViewClient;
 import com.github.dfa.diaspora_android.webview.ProgressBarWebChromeClient;
-import com.github.dfa.diaspora_android.util.AppLog;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -42,16 +57,13 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.Locale;
 
-import info.guardianproject.netcipher.NetCipher;
-import info.guardianproject.netcipher.webkit.WebkitProxy;
-
 /**
  * Fragment with a webView and a ProgressBar.
  * This Fragment retains its instance.
  * Created by vanitas on 26.09.16.
  */
 
-public class BrowserFragment extends CustomFragment {
+public class BrowserFragment extends ThemedFragment {
     public static final String TAG = "com.github.dfa.diaspora_android.BrowserFragment";
 
     protected View rootLayout;
@@ -84,19 +96,11 @@ public class BrowserFragment extends CustomFragment {
         if(this.webView == null) {
             this.webView = (ContextMenuWebView) view.findViewById(R.id.webView);
             this.applyWebViewSettings();
+            ProxyHandler.getInstance().addWebView(webView);
         }
 
         if(this.progressBar == null) {
             this.progressBar = (ProgressBar) view.findViewById(R.id.progressBar);
-        }
-
-        if (appSettings.isProxyEnabled()) {
-            if (!setProxy(appSettings.getProxyHost(), appSettings.getProxyPort())) {
-                AppLog.e(this, "Could not enable Proxy");
-                Toast.makeText(getContext(), R.string.toast_set_proxy_failed, Toast.LENGTH_SHORT).show();
-            }
-        } else if (appSettings.wasProxyEnabled()) {
-            resetProxy();
         }
 
         if(pendingUrl != null) {
@@ -142,84 +146,13 @@ public class BrowserFragment extends CustomFragment {
         webView.setWebChromeClient(new ProgressBarWebChromeClient(webView, progressBar));
     }
 
-    /**
-     * Set proxy according to arguments. host must not be "" or null, port must be positive.
-     * Return true on success and update appSettings' proxy related values.
-     *
-     * @param host proxy host (eg. localhost or 127.0.0.1)
-     * @param port proxy port (eg. 8118)
-     * @return success
-     * @throws IllegalArgumentException if arguments do not fit specifications above
-     */
-    private boolean setProxy(final String host, final int port) {
-        AppLog.i(this, "StreamFragment.setProxy()");
-        if (host != null && !host.equals("") && port >= 0) {
-            AppLog.i(this, "Set proxy to "+host+":"+port);
-            //Temporary change thread policy
-            AppLog.v(this, "Set temporary ThreadPolicy");
-            StrictMode.ThreadPolicy old = StrictMode.getThreadPolicy();
-            StrictMode.ThreadPolicy tmp = new StrictMode.ThreadPolicy.Builder().permitAll().build();
-            StrictMode.setThreadPolicy(tmp);
-
-            AppLog.v(this, "Apply NetCipher proxy settings");
-            NetCipher.setProxy(host, port); //Proxy for HttpsUrlConnections
-            try {
-                //Proxy for the webview
-                AppLog.v(this, "Apply Webkit proxy settings");
-                WebkitProxy.setProxy(MainActivity.class.getName(), getContext().getApplicationContext(), null, host, port);
-            } catch (Exception e) {
-                AppLog.e(this, "Could not apply WebKit proxy settings:\n"+e.toString());
-            }
-            AppLog.v(this, "Save changes in appSettings");
-            appSettings.setProxyEnabled(true);
-            appSettings.setProxyWasEnabled(true);
-
-            AppLog.v(this, "Reset old ThreadPolicy");
-            StrictMode.setThreadPolicy(old);
-            AppLog.i(this, "Success! Reload WebView");
-            webView.reload();
-            return true;
-        } else {
-            AppLog.e(this, "Invalid proxy configuration. Host: "+host+" Port: "+port+"\nRefuse to set proxy");
-            return false;
+    @Override
+    public void onResume() {
+        super.onResume();
+        if(webView != null) {
+            webSettings.setMinimumFontSize(appSettings.getMinimumFontSize());
+            webSettings.setLoadsImagesAutomatically(appSettings.isLoadImages());
         }
-    }
-
-    @SuppressWarnings("unused")
-    private boolean setProxy() {
-        return setProxy(appSettings.getProxyHost(), appSettings.getProxyPort());
-    }
-
-    private void resetProxy() {
-        AppLog.i(this, "StreamFragment.resetProxy()");
-        AppLog.v(this, "write changes to appSettings");
-        appSettings.setProxyEnabled(false);
-        appSettings.setProxyWasEnabled(false);
-
-        //Temporary change thread policy
-        AppLog.v(this, "Set temporary ThreadPolicy");
-        StrictMode.ThreadPolicy old = StrictMode.getThreadPolicy();
-        StrictMode.ThreadPolicy tmp = new StrictMode.ThreadPolicy.Builder().permitAll().build();
-        StrictMode.setThreadPolicy(tmp);
-
-        AppLog.v(this, "clear NetCipher proxy");
-        NetCipher.clearProxy();
-        try {
-            AppLog.v(this, "clear WebKit proxy");
-            WebkitProxy.resetProxy(MainActivity.class.getName(), getContext());
-        } catch (Exception e) {
-            AppLog.e(this, "Could not clear WebKit proxy:\n"+e.toString());
-        }
-        AppLog.v(this, "Reset old ThreadPolicy");
-        StrictMode.setThreadPolicy(old);
-
-        //Restart app
-        AppLog.i(this, "Success! Restart app due to proxy reset");
-        Intent restartActivity = new Intent(getContext(), MainActivity.class);
-        PendingIntent pendingIntent = PendingIntent.getActivity(getContext(), 12374, restartActivity, PendingIntent.FLAG_CANCEL_CURRENT);
-        AlarmManager mgr = (AlarmManager) getActivity().getSystemService(Context.ALARM_SERVICE);
-        mgr.set(AlarmManager.RTC, System.currentTimeMillis() + 100, pendingIntent);
-        System.exit(0);
     }
 
     @SuppressWarnings("ResultOfMethodCallIgnored")
@@ -349,5 +282,10 @@ public class BrowserFragment extends CustomFragment {
 
     public ContextMenuWebView getWebView() {
         return this.webView;
+    }
+
+    @Override
+    protected void applyColorToViews() {
+        ThemeHelper.updateProgressBarColor(progressBar);
     }
 }